+++ src/commands/start.ts
@@ -16,7 +16,7 @@
 
 import { Command } from '@oclif/command';
 import { LoggerFactory } from '../logger';
-import { BootstrapService, CommandUtils } from '../service';
+import { BootstrapService, BootstrapUtils, CommandUtils } from '../service';
 import Clean from './clean';
 import Compose from './compose';
 import Config from './config';
@@ -30,6 +30,8 @@ export default class Start extends Command {
         `$ symbol-bootstrap start -p testnet -a dual`,
         `$ symbol-bootstrap start -p mainnet -a peer -c custom-preset.yml`,
         `$ symbol-bootstrap start -p testnet -a dual --password 1234`,
+        `$ symbol-bootstrap start -p mainnet -a my-custom-assembly.yml -c custom-preset.yml`,
+        `$ symbol-bootstrap start -p my-custom-network.yml -a dual -c custom-preset.yml`,
         `$ echo "$MY_ENV_VAR_PASSWORD" | symbol-bootstrap start -p testnet -a dual`,
     ];
 
@@ -46,6 +48,6 @@ export default class Start extends Command {
             CommandUtils.passwordPromptDefaultMessage,
             true,
         );
-        await new BootstrapService(logger).start(flags);
+        await new BootstrapService(logger).start({ ...flags, workingDir: BootstrapUtils.defaultWorkingDir });
     }
 }