+++ test/service/ConfigLoader.test.ts
@@ -38,6 +38,7 @@ describe('ConfigLoader', () => {
         const configLoader = new ConfigLoaderMocked(logger);
         try {
             await configLoader.createPresetData({
+                workingDir: BootstrapUtils.defaultWorkingDir,
                 preset: undefined,
                 assembly: undefined,
                 customPreset: undefined,
@@ -56,6 +57,7 @@ describe('ConfigLoader', () => {
         const configLoader = new ConfigLoaderMocked(logger);
         try {
             await configLoader.createPresetData({
+                workingDir: BootstrapUtils.defaultWorkingDir,
                 preset: Preset.testnet,
                 assembly: undefined,
                 customPreset: undefined,
@@ -70,9 +72,46 @@ describe('ConfigLoader', () => {
         }
     });
 
+    it('ConfigLoader loadPresetData invalid assembly file', async () => {
+        const configLoader = new ConfigLoaderMocked(logger);
+        try {
+            await configLoader.createPresetData({
+                workingDir: BootstrapUtils.defaultWorkingDir,
+                preset: Preset.testnet,
+                assembly: 'invalid-assembly.yml',
+                customPreset: undefined,
+                customPresetObject: undefined,
+                password: 'abc',
+            });
+            expect(false).to.be.eq(true); // should have raised an error!
+        } catch (e) {
+            expect(e.message).eq(
+                "Assembly 'invalid-assembly.yml' is not valid for preset 'testnet'. Have you provided the right --preset <preset> --assembly <assembly> ?",
+            );
+        }
+    });
+
+    it('ConfigLoader loadPresetData invalid preset file', async () => {
+        const configLoader = new ConfigLoaderMocked(logger);
+        try {
+            await configLoader.createPresetData({
+                workingDir: BootstrapUtils.defaultWorkingDir,
+                preset: 'invalid-preset.yml',
+                assembly: Assembly.dual,
+                customPreset: undefined,
+                customPresetObject: undefined,
+                password: 'abc',
+            });
+            expect(false).to.be.eq(true); // should have raised an error!
+        } catch (e) {
+            expect(e.message).eq("Preset 'invalid-preset.yml' does not exist. Have you provided the right --preset <preset> ?");
+        }
+    });
+
     it('ConfigLoader loadPresetData bootstrap no assembly', async () => {
         const configLoader = new ConfigLoaderMocked(logger);
         const presetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             preset: Preset.bootstrap,
             customPreset: undefined,
             customPresetObject: undefined,
@@ -86,6 +125,7 @@ describe('ConfigLoader', () => {
     it('ConfigLoader loadPresetData testnet assembly', async () => {
         const configLoader = new ConfigLoaderMocked(logger);
         const presetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             preset: Preset.testnet,
             assembly: Assembly.api,
             customPreset: undefined,
@@ -97,9 +137,25 @@ describe('ConfigLoader', () => {
         expect(presetData.assembly).to.eq(Assembly.api);
     });
 
+    it('ConfigLoader loadPresetData testnet assembly using external files', async () => {
+        const configLoader = new ConfigLoaderMocked(logger);
+        const presetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
+            preset: 'presets/testnet/network.yml',
+            assembly: 'presets/assemblies/assembly-api.yml',
+            customPreset: undefined,
+            customPresetObject: undefined,
+            password: 'abc',
+        });
+        expect(presetData).to.not.be.undefined;
+        expect(presetData.preset).to.eq('presets/testnet/network.yml');
+        expect(presetData.assembly).to.eq('presets/assemblies/assembly-api.yml');
+    });
+
     it('ConfigLoader custom maxUnlockedAccounts', async () => {
         const configLoader = new ConfigLoaderMocked(logger);
         const originalPresetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             preset: Preset.testnet,
             assembly: Assembly.dual,
             customPreset: 'test/unit-test-profiles/custom_preset.yml',
@@ -109,6 +165,7 @@ describe('ConfigLoader', () => {
             password: 'abcd',
         });
         expect(originalPresetData).to.not.be.undefined;
+        expect(originalPresetData.nodes![0].maxUnlockedAccounts).eq(undefined);
         expect(originalPresetData.maxUnlockedAccounts).eq(30);
         expect(originalPresetData.customPresetCache).deep.eq({
             maxUnlockedAccounts: 30,
@@ -126,12 +183,14 @@ describe('ConfigLoader', () => {
         });
 
         const upgradedPresetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             oldPresetData: originalPresetData,
             preset: Preset.testnet,
             assembly: Assembly.dual,
             password: 'abcd',
         });
         expect(upgradedPresetData).to.not.be.undefined;
+        expect(upgradedPresetData.nodes![0].maxUnlockedAccounts).eq(undefined);
         expect(upgradedPresetData.maxUnlockedAccounts).eq(30);
         expect(upgradedPresetData.customPresetCache).deep.eq({
             maxUnlockedAccounts: 30,
@@ -149,6 +208,7 @@ describe('ConfigLoader', () => {
         });
 
         const upgradedPresetResetToDefaults = await configLoader.createPresetData({
+            workingDir: '.',
             oldPresetData: originalPresetData,
             preset: Preset.testnet,
             customPresetObject: {
@@ -158,6 +218,7 @@ describe('ConfigLoader', () => {
             password: 'abcd',
         });
         expect(upgradedPresetResetToDefaults).to.not.be.undefined;
+        expect(upgradedPresetResetToDefaults.nodes![0].maxUnlockedAccounts).eq(undefined);
         expect(upgradedPresetResetToDefaults.maxUnlockedAccounts).eq(15);
         expect(upgradedPresetResetToDefaults.customPresetCache).deep.eq({
             maxUnlockedAccounts: 15,
@@ -204,6 +265,7 @@ describe('ConfigLoader', () => {
     it('ConfigLoader custom maxUnlockedAccounts', async () => {
         const configLoader = new ConfigLoaderMocked(logger);
         const originalPresetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             preset: Preset.testnet,
             assembly: Assembly.dual,
             customPreset: 'test/unit-test-profiles/custom_preset.yml',
@@ -231,6 +293,7 @@ describe('ConfigLoader', () => {
         });
 
         const upgradedPresetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             oldPresetData: originalPresetData,
             preset: Preset.testnet,
             assembly: Assembly.dual,
@@ -255,6 +318,7 @@ describe('ConfigLoader', () => {
         });
 
         const upgradedPresetResetToDefaults = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             oldPresetData: originalPresetData,
             preset: Preset.testnet,
             customPresetObject: {
@@ -311,6 +375,7 @@ describe('ConfigLoader', () => {
     it('ConfigLoader loadPresetData bootstrap custom', async () => {
         const configLoader = new ConfigLoaderMocked(logger);
         const presetData = await configLoader.createPresetData({
+            workingDir: BootstrapUtils.defaultWorkingDir,
             preset: Preset.bootstrap,
             assembly: undefined,
             customPreset: 'test/override-currency-preset.yml',
@@ -327,6 +392,7 @@ describe('ConfigLoader', () => {
         const configLoader = new ConfigLoaderMocked(logger);
         try {
             await configLoader.createPresetData({
+                workingDir: BootstrapUtils.defaultWorkingDir,
                 preset: Preset.bootstrap,
                 assembly: undefined,
                 customPreset: 'test/override-currency-preset.yml',